requirement.txt
asgiref==3.3.4
certifi==2021.5.30
chardet==4.0.0
defusedxml==0.7.1
Django==3.2.4
django-allauth==0.42.0
django-cors-headers==3.7.0
django-crispy-forms==1.11.2
django-filter==2.4.0
django-widget-tweaks==1.4.8
djangorestframework==3.12.4
djangorestframework-filters==0.11.1
gunicorn==20.1.0
idna==2.10
oauthlib==3.1.1
Pillow==8.2.0
pkg-resources==0.0.0
python3-openid==3.2.0
pytz==2021.1
requests==2.25.1
requests-oauthlib==1.3.0
sqlparse==0.4.1
typing-extensions==3.10.0.0
urllib3==1.26.5
whitenoise==5.2.0
